+++ packages/components/src/components/pagination/pagination.stories.ts
@@ -1,16 +1,15 @@
 import { defaultLocale } from "@arcgis/toolkit/intl";
-import { numberingSystems, defaultNumberingSystem } from "../../utils/locale";
-import { supportedNlsLocales } from "../date-picker/utils";
-import { createBreakpointStories, modesDarkDefault } from "../../../.storybook/utils";
+import { defaultNumberingSystem } from "../../utils/locale";
+import { boolean, createBreakpointStories, modesDarkDefault } from "../../../.storybook/utils";
 import { html } from "../../../support/formatting";
 import { ATTRIBUTES } from "../../../.storybook/resources";
 import { Pagination } from "./pagination";
 
-const { scale } = ATTRIBUTES;
+const { numberingSystem, scale, supportedNlsLocale } = ATTRIBUTES;
 
 interface PaginationStoryArgs extends Pick<
   Pagination,
-  "scale" | "startItem" | "numberingSystem" | "totalItems" | "pageSize"
+  "groupSeparator" | "numberingSystem" | "pageSize" | "scale" | "startItem" | "totalItems"
 > {
   lang: string;
 }
@@ -21,7 +20,8 @@ export default {
     scale: scale.defaultValue,
     startItem: 1,
     lang: defaultLocale,
-    numberingSystem: defaultNumberingSystem,
+    numberingSystem: numberingSystem.defaultValue,
+    groupSeparator: false,
     totalItems: 123456789,
     pageSize: 10,
   },
@@ -31,11 +31,11 @@ export default {
       control: { type: "select" },
     },
     lang: {
-      options: supportedNlsLocales,
+      options: supportedNlsLocale.values,
       control: { type: "select" },
     },
     numberingSystem: {
-      options: numberingSystems,
+      options: numberingSystem.values,
       control: { type: "select" },
     },
   },
@@ -59,6 +59,7 @@ export const simple = (args: PaginationStoryArgs): string => html`
     start-item="${args.startItem}"
     lang="${args.lang}"
     numbering-system="${args.numberingSystem}"
+    ${boolean("group-separator", args.groupSeparator)}
     total-items="${args.totalItems}"
     page-size="${args.pageSize}"
   >