+++ packages/components/src/components/slider/slider.stories.ts
@@ -1,17 +1,19 @@
-import { boolean, modesDarkDefault } from "../../../.storybook/utils";
+import { boolean, modesDarkDefault, optionalAttribute } from "../../../.storybook/utils";
 import { html } from "../../../support/formatting";
 import { iconNames } from "../../../.storybook/helpers";
 import { ATTRIBUTES } from "../../../.storybook/resources";
 import type { Slider } from "./slider";
 
-const { scale, status } = ATTRIBUTES;
+const { scale, sliderFillPlacement, status } = ATTRIBUTES;
 
 interface SliderStoryArgs extends Pick<
   Slider,
   | "min"
   | "max"
+  | "maxLabel"
   | "value"
   | "step"
+  | "fillPlacement"
   | "minLabel"
   | "disabled"
   | "labelHandles"
@@ -23,8 +25,8 @@ interface SliderStoryArgs extends Pick<
   | "snap"
   | "scale"
   | "status"
-  | "validationMessage"
   | "validationIcon"
+  | "validationMessage"
 > {
   temperature: string;
 }
@@ -36,6 +38,8 @@ export default {
     max: 100,
     value: 50,
     step: 1,
+    maxLabel: "",
+    fillPlacement: "all",
     minLabel: "Temperature",
     disabled: false,
     labelHandles: false,
@@ -63,6 +67,10 @@ export default {
       options: iconNames,
       control: { type: "select" },
     },
+    fillPlacement: {
+      options: sliderFillPlacement.values,
+      control: { type: "select" },
+    },
   },
   parameters: {
     chromatic: {
@@ -77,8 +85,10 @@ export const simple = (args: SliderStoryArgs): string => html`
   <calcite-slider
     min="${args.min}"
     max="${args.max}"
+    max-label="${args.maxLabel}"
     value="${args.value}"
     step="${args.step}"
+    fill-placement="${args.fillPlacement}"
     min-label="${args.minLabel}"
     ${boolean("disabled", args.disabled)}
     ${boolean("label-handles", args.labelHandles)}
@@ -91,7 +101,7 @@ export const simple = (args: SliderStoryArgs): string => html`
     scale="${args.scale}"
     status="${args.status}"
     validation-message="${args.validationMessage}"
-    validation-icon="${args.validationIcon}"
+    ${optionalAttribute("validation-icon", args.validationIcon)}
   ></calcite-slider>
 `;