+++ packages/components/src/components/swatch/swatch.stories.ts
@@ -7,16 +7,23 @@ import { Swatch } from "./swatch";
 const { scale } = ATTRIBUTES;
 
 type SwatchStoryArgs = Pick<Swatch, "scale" | "selected" | "label">;
+type SwatchSimpleStoryArgs = Pick<Swatch, "color" | "disabled" | "label" | "scale" | "selected">;
 
 export default {
   title: "Components/Swatch",
-  args: { scale: scale.defaultValue, selected: false, label: "My great swatch" },
+  args: { color: "", disabled: false, scale: scale.defaultValue, selected: false, label: "My great swatch" },
   argTypes: { scale: { options: scale.values, control: { type: "select" } }, label: { control: { type: "text" } } },
 };
 
-export const simple = (args: SwatchStoryArgs): string => html`
+export const simple = (args: SwatchSimpleStoryArgs): string => html`
   <div style="background-color:white;padding:100px">
-    <calcite-swatch scale="${args.scale}" label="${args.label}" ${boolean("selected", args.selected)}></calcite-swatch>
+    <calcite-swatch
+      scale="${args.scale}"
+      color="${args.color}"
+      label="${args.label}"
+      ${boolean("disabled", args.disabled)}
+      ${boolean("selected", args.selected)}
+    ></calcite-swatch>
   </div>
 `;